Mimic previous behavior and buffer rendering of non-chunked bodies until we complete deferred output.

+++ src/main/java/com/vtence/molecule/Body.java
@@ -6,7 +6,7 @@ import java.io.OutputStream;
 
 public interface Body extends Closeable {
 
-    long size();
+    int size();
 
     void writeTo(OutputStream out) throws IOException;
 }
